backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
autobahn==23.6.2
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.14.3
billiard==4.2.3
bleach==6.1.0
celery==5.3.0
certifi==2025.11.12
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
constantly==23.10.4
coverage==7.5.4
cryptography==46.0.3
daphne==4.1.2
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.8
django-cors-headers==4.9.0
django-reverse-js==0.1.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
iniconfig==2.3.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.0
lark==1.3.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
matplotlib==3.10.6
mistune==3.1.4
nest-asyncio==1.6.0
numpy==2.1.3
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyOpenSSL==25.3.0
pytest==9.0.1
pytest-django==4.9.0
psycopg[binary]==3.2.*
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
zope.interface==8.1.1
Pillow>=10.0.0
